// src/app/constants/item-type.enum.ts
export enum ItemTypeEnum {
  weapon = 'weapon',
  weaponGrade = 'weaponGrade',
  weaponRefine = 'weaponRefine',
  weaponCard1 = 'weaponCard1',
  weaponCard2 = 'weaponCard2',
  weaponCard3 = 'weaponCard3',
  weaponCard4 = 'weaponCard4',
  weaponEnchant0 = 'weaponEnchant0',
  weaponEnchant1 = 'weaponEnchant1',
  weaponEnchant2 = 'weaponEnchant2',
  weaponEnchant3 = 'weaponEnchant3',

  ammo = 'ammo',

  leftWeapon = 'leftWeapon',
  leftWeaponRefine = 'leftWeaponRefine',
  leftWeaponGrade = 'leftWeaponGrade',
  leftWeaponCard1 = 'leftWeaponCard1',
  leftWeaponCard2 = 'leftWeaponCard2',
  leftWeaponCard3 = 'leftWeaponCard3',
  leftWeaponCard4 = 'leftWeaponCard4',
  leftWeaponEnchant0 = 'leftWeaponEnchant0',
  leftWeaponEnchant1 = 'leftWeaponEnchant1',
  leftWeaponEnchant2 = 'leftWeaponEnchant2',
  leftWeaponEnchant3 = 'leftWeaponEnchant3',

  shield = 'shield',
  shieldRefine = 'shieldRefine',
  shieldGrade = 'shieldGrade',
  shieldCard = 'shieldCard',
  shieldEnchant1 = 'shieldEnchant1',
  shieldEnchant2 = 'shieldEnchant2',
  shieldEnchant3 = 'shieldEnchant3',

  headUpper = 'headUpper',
  headUpperRefine = 'headUpperRefine',
  headUpperGrade = 'headUpperGrade',
  headUpperCard = 'headUpperCard',
  headUpperEnchant1 = 'headUpperEnchant1',
  headUpperEnchant2 = 'headUpperEnchant2',
  headUpperEnchant3 = 'headUpperEnchant3',
  headMiddle = 'headMiddle',
  headMiddleCard = 'headMiddleCard',
  headMiddleGrade = 'headMiddleGrade',
  headMiddleEnchant1 = 'headMiddleEnchant1',
  headMiddleEnchant2 = 'headMiddleEnchant2',
  headMiddleEnchant3 = 'headMiddleEnchant3',
  headLower = 'headLower',
  headLowerCard = 'headLowerCard',
  headLowerGrade = 'headLowerGrade',
  headLowerEnchant1 = 'headLowerEnchant1',
  headLowerEnchant2 = 'headLowerEnchant2',
  headLowerEnchant3 = 'headLowerEnchant3',
  armor = 'armor',
  armorRefine = 'armorRefine',
  armorGrade = 'armorGrade',
  armorCard = 'armorCard',
  armorEnchant1 = 'armorEnchant1',
  armorEnchant2 = 'armorEnchant2',
  armorEnchant3 = 'armorEnchant3',
  garment = 'garment',
  garmentRefine = 'garmentRefine',
  garmentGrade = 'garmentGrade',
  garmentCard = 'garmentCard',
  garmentEnchant1 = 'garmentEnchant1',
  garmentEnchant2 = 'garmentEnchant2',
  garmentEnchant3 = 'garmentEnchant3',
  boot = 'boot',
  bootRefine = 'bootRefine',
  bootGrade = 'bootGrade',
  bootCard = 'bootCard',
  bootEnchant1 = 'bootEnchant1',
  bootEnchant2 = 'bootEnchant2',
  bootEnchant3 = 'bootEnchant3',
  accLeft = 'accLeft',
  accLeftRefine = 'accLeftRefine',
  accLeftGrade = 'accLeftGrade',
  accLeftCard = 'accLeftCard',
  accLeftEnchant1 = 'accLeftEnchant1',
  accLeftEnchant2 = 'accLeftEnchant2',
  accLeftEnchant3 = 'accLeftEnchant3',
  accRight = 'accRight',
  accRightRefine = 'accRightRefine',
  accRightGrade = 'accRightGrade',
  accRightCard = 'accRightCard',
  accRightEnchant1 = 'accRightEnchant1',
  accRightEnchant2 = 'accRightEnchant2',
  accRightEnchant3 = 'accRightEnchant3',
  pet = 'pet',

  costumeUpper = 'costumeUpper',
  costumeMiddle = 'costumeMiddle',
  costumeLower = 'costumeLower',
  costumeGarment = 'costumeGarment',

  costumeEnchantUpper = 'costumeEnchantUpper',
  costumeEnchantMiddle = 'costumeEnchantMiddle',
  costumeEnchantLower = 'costumeEnchantLower',
  costumeEnchantGarment = 'costumeEnchantGarment',
  costumeEnchantGarment2 = 'costumeEnchantGarment2',
  costumeEnchantGarment4 = 'costumeEnchantGarment4',

  shadowWeapon = 'shadowWeapon',
  shadowWeaponRefine = 'shadowWeaponRefine',
  shadowWeaponEnchant2 = 'shadowWeaponEnchant2',
  shadowWeaponEnchant3 = 'shadowWeaponEnchant3',
  shadowArmor = 'shadowArmor',
  shadowArmorRefine = 'shadowArmorRefine',
  shadowArmorEnchant2 = 'shadowArmorEnchant2',
  shadowArmorEnchant3 = 'shadowArmorEnchant3',
  shadowShield = 'shadowShield',
  shadowShieldRefine = 'shadowShieldRefine',
  shadowShieldEnchant2 = 'shadowShieldEnchant2',
  shadowShieldEnchant3 = 'shadowShieldEnchant3',
  shadowBoot = 'shadowBoot',
  shadowBootRefine = 'shadowBootRefine',
  shadowBootEnchant2 = 'shadowBootEnchant2',
  shadowBootEnchant3 = 'shadowBootEnchant3',
  shadowEarring = 'shadowEarring',
  shadowEarringRefine = 'shadowEarringRefine',
  shadowEarringEnchant2 = 'shadowEarringEnchant2',
  shadowEarringEnchant3 = 'shadowEarringEnchant3',
  shadowPendant = 'shadowPendant',
  shadowPendantRefine = 'shadowPendantRefine',
  shadowPendantEnchant2 = 'shadowPendantEnchant2',
  shadowPendantEnchant3 = 'shadowPendantEnchant3',
}